map refacror

+++ Entity/Map/utils.h
@@ -2,5 +2,5 @@
 #define UTILS_H_
 #include <random>
 
-int randNum(long long a, long long b);
+int randNum(int a, int b);
 #endif
